Move SnapshotJsonRpcClient up
